fix: 页表项用 unsigned long long —— Windows 上 unsigned long 只有 32 位
跨平台 CI 抓到的:

    tests/pte_encoding.cpp:117: error: constexpr variable 'kPxn' must be
    initialized by a constant expression

一个页表项在两台机器上都是 64 位,而 `unsigned long` 在写这段代码的系统上
恰好也是 64 位。它在 Windows 上是 32 位,于是 `1UL << 53` 是宽于类型的移位
—— 未定义,而且实践中是**静默取零**而不是报错。在那里构建出的编码器会产出
高位字段全部缺失的表项,并且没有任何诊断。

宿主作业之所以能抓到,是因为 `constexpr` 强制在编译期求值,把静默的那种
情形变成了「must be initialized by a constant expression」。换句话说:
**如果这些常量不是 constexpr,这个缺陷会一路静默地进到镜像里。**

`unsigned long long` 在任何地方都至少 64 位,这正是表项需要的性质。

上下文那一侧的 `unsigned long` 不动:那些文件只为裸机目标编译,两个目标
都是 LP64。

+++ src/arch/aarch64/pte_impl.cpp
@@ -5,13 +5,13 @@ module openarch.pte;
 
 namespace arch {
 
-pte make_leaf(unsigned long phys, perm p, memory_type mt, bool user) noexcept {
+pte make_leaf(unsigned long long phys, perm p, memory_type mt, bool user) noexcept {
     return pte{ aarch64::encode_leaf(phys, static_cast<int>(p),
                                      static_cast<int>(mt), user) };
 }
 
 bool          is_valid(pte e) noexcept { return aarch64::entry_valid(e.bits); }
-unsigned long phys_of (pte e) noexcept { return aarch64::entry_phys(e.bits);  }
+unsigned long long phys_of (pte e) noexcept { return aarch64::entry_phys(e.bits);  }
 
 // The canonical MAIR_EL1 layout the encoder's AttrIndx values index into.
 //
@@ -31,7 +31,7 @@ unsigned long phys_of (pte e) noexcept { return aarch64::entry_phys(e.bits);  }
 // mappings installed after boot may be walked against the reset value of the
 // register, which is architecturally UNKNOWN.
 void install_memory_attributes() noexcept {
-    constexpr unsigned long kMair = 0x00FFUL;
+    constexpr unsigned long long kMair = 0x00FFUL;
     asm volatile("msr mair_el1, %0\n\tisb" :: "r"(kMair) : "memory");
 }
